+++ api/authentication/tests/__init__.py
@@ -1,1 +1,2 @@
-from .test import *
+from .test_models import *
+from .test_views import *
